refactor: 북마크 기능 패키지 생성
body:
