+++ src/main/java/com/team_7/moment_film/domain/bookmark/controller/FrameBookmarkController.java
@@ -1,0 +1,30 @@
+package com.team_7.moment_film.domain.bookmark.controller;
+
+import com.team_7.moment_film.domain.bookmark.service.FrameBookmarkService;
+import com.team_7.moment_film.global.dto.ApiResponse;
+import com.team_7.moment_film.global.security.UserDetailsImpl;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/api/frame")
+@RequiredArgsConstructor
+public class FrameBookmarkController {
+
+    private final FrameBookmarkService bookMarkService;
+
+    // 북마크 추가/취소 기능
+    @PostMapping("/{frameId}/bookmark")
+    public ResponseEntity<ApiResponse> BookMarkFilter(@PathVariable Long frameId, @AuthenticationPrincipal UserDetailsImpl userDetails){
+        return bookMarkService.bookMarkFrame(frameId, userDetails.getUser());
+    }
+
+    // 북마크한 프레임 조회 기능
+    @GetMapping("/bookmark")
+    public ResponseEntity<ApiResponse> GetBookMarkFilter(@AuthenticationPrincipal UserDetailsImpl userDetails){
+        return bookMarkService.getBookMarkFrame(userDetails.getUser());
+    }
+
+}
