+++ src/main/java/com/team_7/moment_film/domain/bookmark/service/FilterBookmarkService.java
@@ -1,9 +1,9 @@
-package com.team_7.moment_film.domain.customfilter.service;
+package com.team_7.moment_film.domain.bookmark.service;
 
 import com.team_7.moment_film.domain.customfilter.dto.FilterResponseDto;
 import com.team_7.moment_film.domain.customfilter.entity.Filter;
-import com.team_7.moment_film.domain.customfilter.entity.FilterBookMark;
-import com.team_7.moment_film.domain.customfilter.repository.FilterBookMarkRepository;
+import com.team_7.moment_film.domain.bookmark.entity.FilterBookmark;
+import com.team_7.moment_film.domain.bookmark.repository.FilterBookmarkRepository;
 import com.team_7.moment_film.domain.customfilter.repository.FilterRepository;
 import com.team_7.moment_film.domain.user.entity.User;
 import com.team_7.moment_film.global.dto.ApiResponse;
@@ -18,9 +18,9 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class FilterBookMarkService {
+public class FilterBookmarkService {
 
-    private final FilterBookMarkRepository bookMarkRepository;
+    private final FilterBookmarkRepository bookMarkRepository;
     private final FilterRepository filterRepository;
 
     // 필터 북마크 추가/취소
@@ -32,7 +32,7 @@ public class FilterBookMarkService {
 
         // 북마크 없으면 추가
         if(!bookmarked){
-            FilterBookMark bookMark = FilterBookMark.builder().user(user).filter(filter).build();
+            FilterBookmark bookMark = FilterBookmark.builder().user(user).filter(filter).build();
             bookMarkRepository.save(bookMark);
 
             ApiResponse apiResponse = ApiResponse.builder().status(HttpStatus.OK).msg("북마크 완료").build();
