+++ src/main/java/com/team_7/moment_film/domain/customfilter/controller/FilterController.java
@@ -1,7 +1,6 @@
 package com.team_7.moment_film.domain.customfilter.controller;
 
 import com.team_7.moment_film.domain.customfilter.dto.FilterRequestDto;
-import com.team_7.moment_film.domain.customfilter.service.FilterBookMarkService;
 import com.team_7.moment_film.domain.customfilter.service.FilterService;
 import com.team_7.moment_film.global.dto.ApiResponse;
 import com.team_7.moment_film.global.security.UserDetailsImpl;
@@ -17,7 +16,6 @@ import org.springframework.web.bind.annotation.*;
 public class FilterController {
 
     private final FilterService filterService;
-    private final FilterBookMarkService bookMarkService;
 
     // 필터 커스텀(등록) 기능
     @PostMapping("")
@@ -41,18 +39,6 @@ public class FilterController {
     @DeleteMapping("/{filterId}")
     public ResponseEntity<ApiResponse> deleteFilter(@PathVariable Long filterId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
         return filterService.deleteFilter(filterId, userDetails.getUser());
-    }
-
-    // 북마크 추가/취소 기능
-    @PostMapping("/{filterId}/bookmark")
-    public ResponseEntity<ApiResponse> BookMarkFilter(@PathVariable Long filterId, @AuthenticationPrincipal UserDetailsImpl userDetails){
-        return bookMarkService.bookMarkFilter(filterId, userDetails.getUser());
-    }
-
-    // 북마크한 필터 조회 기능
-    @GetMapping("/bookmark")
-    public ResponseEntity<ApiResponse> GetBookMarkFilter(@AuthenticationPrincipal UserDetailsImpl userDetails){
-        return bookMarkService.getBookMarkFilter(userDetails.getUser());
     }
 
 }
